refactor: use AmascutTracker for pet probability
